requirements.txt
aioredis==1.3.1
amqp==5.0.2
argon2-cffi==20.1.0
asgiref==3.3.1
astroid==2.4.2
async-generator==1.10
async-timeout==3.0.1
attrs==20.3.0
autobahn==20.12.2
Automat==20.2.0
autopep8==1.5.4
backcall==0.2.0
beautifulsoup4==4.9.3
billiard==3.6.3.0
bleach==3.2.1
Brotli==1.0.9
cffi==1.14.4
channels==3.0.2
channels-redis==3.2.0
click==7.1.2
click-didyoumean==0.0.3
click-repl==0.1.6
cmdstanpy==0.9.5
colorlover==0.3.0
constantly==15.1.0
convertdate==2.2.0
cryptography==3.3.1
cufflinks==0.17.3
cycler==0.10.0
Cython==0.29.21
daphne==3.0.1
dash==1.18.1
dash-bootstrap-components==0.11.0
dash-core-components==1.14.1
dash-html-components==1.1.1
dash-renderer==1.8.3
dash-table==4.11.1
decorator==4.4.2
defusedxml==0.6.0
diff-match-patch==20200713
Django==3.0.2
django-appconf==1.0.4
django-author==1.0.2
django-bootstrap4==2.3.1
django-cors-headers==3.6.0
django-crispy-forms==1.10.0
django-debug-toolbar==3.2
django-extensions==3.1.0
django-extra-views==0.13.0
django-filter==2.4.0
django-import-export==2.4.0
django-js-asset==1.2.2
django-model-utils==4.1.1
django-mptt==0.11.0
django-notifications-hq==1.6.0
django-pandas==0.6.2
django-plotly-dash==1.5.0
django-redis==4.12.1
django-select2==7.5.0
django-tables2==2.3.3
django-widget-tweaks==1.4.8
dpd-components==0.1.0
entrypoints==0.3
ephem==3.7.7.1
et-xmlfile==1.0.1
fbprophet==0.7.1
Flask==1.1.2
Flask-Compress==1.8.0
future==0.18.2
gunicorn==20.0.4
hijri-converter==2.1.1
hiredis==1.1.0
holidays==0.10.4
hyperlink==20.0.1
idna==2.10
importlib-metadata==2.1.1
incremental==17.5.0
ipykernel==5.4.2
ipython==7.16.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==5.6.4
itsdangerous==1.1.0
jdcal==1.4.1
jedi==0.17.2
Jinja2==2.11.2
jsonfield==3.1.0
jsonschema==3.2.0
jupyter-client==6.1.7
jupyter-core==4.7.0
jupyterlab-pygments==0.1.2
kiwisolver==1.3.1
kombu==5.0.2
korean-lunar-calendar==0.2.1
lazy-object-proxy==1.4.3
LunarCalendar==0.0.9
MarkupPy==1.14
MarkupSafe==1.1.1
matplotlib==3.3.3
mccabe==0.6.1
mistune==0.8.4
msgpack==1.0.1
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.0.8
nest-asyncio==1.4.3
notebook==6.1.5
numpy==1.19.4
odfpy==1.4.1
openpyxl==3.0.5
packaging==20.8
pandas==1.1.5
pandocfilters==1.4.3
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.0.1
plotly==4.14.1
prometheus-client==0.9.0
prompt-toolkit==3.0.8
psycopg2==2.8.6
psycopg2-binary==2.8.6
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.20
Pygments==2.7.3
pygraphviz==1.6
PyHamcrest==2.0.2
pylint==2.6.0
PyMeeus==0.3.7
pyOpenSSL==20.0.1
pyparsing==2.4.7
pyrsistent==0.17.3
pystan==2.19.1.1
python-dateutil==2.8.1
python-decouple==3.3
pytz==2019.3
PyYAML==5.3.1
pyzmq==20.0.0
redis==3.5.3
retrying==1.3.3
Send2Trash==1.5.0
service-identity==18.1.0
setuptools-git==1.2
six==1.15.0
soupsieve==2.1
sqlparse==0.4.1
swapper==1.1.2.post1
tablib==3.0.0
terminado==0.9.1
testpath==0.4.4
toml==0.10.1
tornado==6.1
tqdm==4.54.1
traitlets==4.3.3
Twisted==20.3.0
txaio==20.4.1
typed-ast==1.4.1
vine==5.0.0
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
xlrd==2.0.1
xlwt==1.3.0
zipp==3.4.0
zope.interface==5.2.0
